+++ YARG.Core/NewParsing/Sync/Tempo2.cs
@@ -5,21 +5,21 @@
     /// </summary>
     public struct Tempo2
     {
-        public const int MICROS_PER_SECOND =  1000000;
+        public const long MICROS_PER_SECOND =  1000000;
         /// <summary>
         /// A factor used to convert to and from BPM and MicrosPerQuarter.<br></br>
         /// There is quite literally no unit of measurement for it.
         /// </summary>
-        public const int BPM_FACTOR = 60000000;
-        public const int DEFAULT_BPM = 120;
-        public const int MICROS_AT_120BPM = BPM_FACTOR / DEFAULT_BPM;
+        public const long BPM_FACTOR = 60000000;
+        public const long DEFAULT_BPM = 120;
+        public const long MICROS_AT_120BPM = BPM_FACTOR / DEFAULT_BPM;
 
         public static readonly Tempo2 DEFAULT = new()
         {
             MicrosecondsPerQuarter = MICROS_AT_120BPM
         };
 
-        public int MicrosecondsPerQuarter;
+        public long MicrosecondsPerQuarter;
         public long PositionInMicroseconds;
 
         /// <summary>
